food_list reviewScore

+++ final_project/src/main/java/model/AvgReviewScore.java
@@ -5,25 +5,31 @@ import java.io.Serializable;
 @SuppressWarnings("serial")
 public class AvgReviewScore implements Serializable{
 	private int foodNo;
-	private double avgReviewScore;
+	private int avgReviewScore;
 	
 	public AvgReviewScore() {
 	}
-	public AvgReviewScore(int foodNo, double avgReviewScore) {
+
+	public AvgReviewScore(int foodNo, int avgReviewScore) {
 		this.foodNo = foodNo;
 		this.avgReviewScore = avgReviewScore;
 	}
+
 	public int getFoodNo() {
 		return foodNo;
 	}
+
 	public void setFoodNo(int foodNo) {
 		this.foodNo = foodNo;
 	}
-	public double getAvgReviewScore() {
+
+	public int getAvgReviewScore() {
 		return avgReviewScore;
 	}
-	public void setAvgReviewScore(double avgReviewScore) {
+
+	public void setAvgReviewScore(int avgReviewScore) {
 		this.avgReviewScore = avgReviewScore;
 	}
+	
 	
 }
